Saving work

+++ abipy/benchmarks/__init__.py
@@ -92,7 +92,7 @@ def bench_main(main):
 
         parser.add_argument("--min-ncpus", default=-1, type=int, help="Minimum number of CPUs to be tested.")
         parser.add_argument("--max-ncpus", default=206, type=int, help="Maximum number of CPUs to be tested. Default: 206.")
-        parser.add_argument("--min-eff", default=0.6, type=int, help="Minimum parallel efficiency accepted. Default 0.6.")
+        parser.add_argument("--min-eff", default=0.6, type=float, help="Minimum parallel efficiency accepted. Default 0.6.")
 
         parser.add_argument('--paw', default=False, action="store_true", help="Run PAW calculation if available")
         parser.add_argument('--validate', default=False, action="store_true", help="Validate input files and return")
